Replace Template Placeholders with Values

The skeleton.Rmd template files contain placeholders which are to be replaced by the values given in the list pl_repl_values.

sub_pattern_replacement(ps_path, pl_repl_values)

Arguments

ps_path
path to new Rmarkdown file

pl_repl_values
list with replacement values

Examples

if (FALSE) {
tmpdir <- tempdir()
s_skeleton_path <- system.file("rmarkdown", "templates", "qemptydoc", "skeleton", "skeleton.Rmd", package = "qrmdtmpl")
s_skeleton_file <- basename(s_skeleton_path)
file.copy(from = s_skeleton_path, to = tmpdir)
s_trg_path <- file.path(tmpdir, s_skeleton_file)
qrmdtmpl:::sub_pattern_replacement(ps_path = s_trg_path,
                        pl_repl_values = list(author = "Peter von Rohr",
                                              title  = "A New Rmarkdown Document",
                                              output_format = "html_document"))
vec_doc <- readLines(con = file(s_trg_path))
cat(paste0(vec_doc, collapse = "\n"), "\n")
file.remove(s_trg_path)
unlink(tmpdir)
}